% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-data_source.R
\name{data_source}
\alias{data_source}
\title{climateR Catalog Data Source}
\value{
(\code{raw()})\cr
The data member of this data source in Arrow IPC Stream format

(\code{character(1)})\cr
The given path, invisibly.
}
\description{
This is a generic base class for describing
catalog data sources. Each data source must
provide a \code{pull} and \code{tidy} method.
}
\note{
If all parameters are missing, then
an empty data_source is created. This is
only useful for reading from IPC.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.id}}{(\code{character(1)})}

\item{\code{.data}}{(\code{arrow::Table})}

\item{\code{.pull}}{(\code{function})}

\item{\code{.tidy}}{(\code{function})}

\item{\code{.finished}}{(\code{logical(1)})\cr
\code{TRUE} if \verb{$tidy()} has been called
successfully, otherwise \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of this data source.}

\item{\code{result}}{(\code{arrow::Table})\cr
Result of this data source after \verb{$tidy()} is called.
(\strong{Read-only})}

\item{\code{.id}}{(\code{character(1)})}

\item{\code{.data}}{(\code{arrow::Table})}

\item{\code{.pull}}{(\code{function})}

\item{\code{.tidy}}{(\code{function})}

\item{\code{.finished}}{(\code{logical(1)})\cr
\code{TRUE} if \verb{$tidy()} has been called
successfully, otherwise \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-data_source-new}{\code{data_source$new()}}
\item \href{#method-data_source-print}{\code{data_source$print()}}
\item \href{#method-data_source-pull}{\code{data_source$pull()}}
\item \href{#method-data_source-tidy}{\code{data_source$tidy()}}
\item \href{#method-data_source-to_ipc_stream}{\code{data_source$to_ipc_stream()}}
\item \href{#method-data_source-from_ipc_stream}{\code{data_source$from_ipc_stream()}}
\item \href{#method-data_source-to_ipc_file}{\code{data_source$to_ipc_file()}}
\item \href{#method-data_source-from_ipc_file}{\code{data_source$from_ipc_file()}}
\item \href{#method-data_source-clone}{\code{data_source$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-new"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-new}{}}}
\subsection{Method \code{new()}}{
Create a new catalog data source.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$new(id, pull, tidy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier for this data source.}

\item{\code{pull}}{(\code{function})\cr
Pull method for this class.\cr

The pull function may require any amount of parameters.\cr

The pull function must return one of:
\code{arrow::Table}, \code{data.table::data.table}, or a \code{data.frame}.}

\item{\code{tidy}}{(\code{function})\cr
Tidy method for this class. See the \code{tidy} method for details.\cr

The tidy function must require at least 1 argument as its
first argument that takes in, and also \strong{returns}, one of:
\code{arrow::Table}, \code{data.table::data.table}, or a \code{data.frame}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-print"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-pull"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-pull}{}}}
\subsection{Method \code{pull()}}{
Pull a catalog data source from its endpoint.

This method is user-defined at object creation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$pull(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{any})\cr
User-defined parameters that may be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-tidy"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-tidy}{}}}
\subsection{Method \code{tidy()}}{
Tidy a raw catalog data source into the catalog schema.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$tidy(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{any})\cr
User-defined parameters that may be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-to_ipc_stream"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-to_ipc_stream}{}}}
\subsection{Method \code{to_ipc_stream()}}{
Marshal this data source to Arrow IPC Stream format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$to_ipc_stream()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-from_ipc_stream"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-from_ipc_stream}{}}}
\subsection{Method \code{from_ipc_stream()}}{
Unmarshals an Arrow IPC Stream to a data source
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$from_ipc_stream(stream)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stream}}{(\code{raw()})\cr
The given Arrow IPC Stream}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-to_ipc_file"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-to_ipc_file}{}}}
\subsection{Method \code{to_ipc_file()}}{
Output this data source to Arrow IPC File format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$to_ipc_file(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{character(1)})\cr
Path to file to write to. Should have extension '.arrow'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-from_ipc_file"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-from_ipc_file}{}}}
\subsection{Method \code{from_ipc_file()}}{
Read a data source from Arrow IPC File format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$from_ipc_file(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{character(1)})\cr
Path to Arrow IPC file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_source-clone"></a>}}
\if{latex}{\out{\hypertarget{method-data_source-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_source$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
