% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data_source.R
\name{new_data_source}
\alias{new_data_source}
\title{Create a new climateR.catalogs data source plugin}
\usage{
new_data_source(name, dir)
}
\arguments{
\item{name}{Name of the data source}

\item{dir}{Directory to output}
}
\description{
Create a new climateR.catalogs data source plugin
}
