% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_stac_children}
\alias{read_stac_children}
\title{Read all child collections from a parent STAC collection}
\usage{
read_stac_children(parent_url, id)
}
\arguments{
\item{parent_url}{Full URL to the parent STAC collection}

\item{id}{Catalog identifier for this data source}
}
\value{
data.frame with rows from all child collections combined
}
\description{
Read all child collections from a parent STAC collection
}
